Move State into Control.Algebra.Internal.
